Refactoring based on review
